requirements.txt
Flask
request
pyjwt